+++ crates/ruma-events/src/call/member.rs
@@ -13,7 +13,7 @@ use std::time::Duration;
 pub use focus::*;
 pub use member_data::*;
 pub use member_state_key::*;
-use ruma_common::{MilliSecondsSinceUnixEpoch, OwnedDeviceId, room_version_rules::RedactionRules};
+use ruma_common::{DeviceId, MilliSecondsSinceUnixEpoch, room_version_rules::RedactionRules};
 use ruma_macros::{EventContent, StringEnum};
 use serde::{Deserialize, Serialize};
 
@@ -68,7 +68,7 @@ impl CallMemberEventContent {
     /// * `expires` - The time after which the event is considered as expired. Defaults to 4 hours.
     pub fn new(
         application: Application,
-        device_id: OwnedDeviceId,
+        device_id: DeviceId,
         focus_active: ActiveFocus,
         foci_preferred: Vec<Focus>,
         created_ts: Option<MilliSecondsSinceUnixEpoch>,
@@ -255,8 +255,7 @@ mod tests {
     use assert_matches2::assert_matches;
     use js_int::{int, uint};
     use ruma_common::{
-        MilliSecondsSinceUnixEpoch, OwnedEventId, OwnedRoomId, OwnedUserId, device_id,
-        owned_device_id, user_id,
+        MilliSecondsSinceUnixEpoch, OwnedEventId, OwnedRoomId, OwnedUserId, device_id, user_id,
     };
     use serde_json::{Value as JsonValue, from_value as from_json_value, json};
     use strass::assert_let;
@@ -280,7 +279,7 @@ mod tests {
                 "123456".to_owned(),
                 CallScope::Room,
             )),
-            device_id: owned_device_id!("ABCDE"),
+            device_id: device_id!("ABCDE"),
             expires: Duration::from_secs(3600),
             foci_active: vec![Focus::Livekit(LivekitFocus {
                 alias: "1".to_owned(),
@@ -294,7 +293,7 @@ mod tests {
     fn create_call_member_event_content() -> CallMemberEventContent {
         CallMemberEventContent::new(
             Application::Call(CallApplicationContent::new("123456".to_owned(), CallScope::Room)),
-            owned_device_id!("ABCDE"),
+            device_id!("ABCDE"),
             ActiveFocus::Livekit(ActiveLivekitFocus {
                 focus_selection: FocusSelection::OldestMembership,
             }),
@@ -373,7 +372,7 @@ mod tests {
     fn deserialize_call_member_event_content() {
         let call_member_ev = CallMemberEventContent::new(
             Application::Call(CallApplicationContent::new("123456".to_owned(), CallScope::Room)),
-            owned_device_id!("THIS_DEVICE"),
+            device_id!("THIS_DEVICE"),
             ActiveFocus::Livekit(ActiveLivekitFocus {
                 focus_selection: FocusSelection::OldestMembership,
             }),
@@ -426,7 +425,7 @@ mod tests {
                 scope: CallScope::Room,
                 call_intent: Some(CallIntent::Audio),
             }),
-            owned_device_id!("THIS_DEVICE"),
+            device_id!("THIS_DEVICE"),
             ActiveFocus::Livekit(ActiveLivekitFocus {
                 focus_selection: FocusSelection::OldestMembership,
             }),
@@ -526,7 +525,7 @@ mod tests {
                     "123456".to_owned(),
                     CallScope::Room,
                 )),
-                device_id: owned_device_id!("THIS_DEVICE"),
+                device_id: device_id!("THIS_DEVICE"),
                 expires: Duration::from_secs(3600),
                 foci_active: vec![Focus::Livekit(LivekitFocus {
                     alias: "room1".to_owned(),
@@ -540,7 +539,7 @@ mod tests {
                     "".to_owned(),
                     CallScope::Room,
                 )),
-                device_id: owned_device_id!("OTHER_DEVICE"),
+                device_id: device_id!("OTHER_DEVICE"),
                 expires: Duration::from_secs(3600),
                 foci_active: vec![Focus::Livekit(LivekitFocus {
                     alias: "room2".to_owned(),
@@ -648,7 +647,7 @@ mod tests {
                 "".to_owned(),
                 CallScope::Room,
             )),
-            device_id: owned_device_id!("THIS_DEVICE"),
+            device_id: device_id!("THIS_DEVICE"),
             foci_preferred: [Focus::Livekit(LivekitFocus {
                 alias: "room1".to_owned(),
                 service_url: "https://livekit1.com".to_owned(),
@@ -775,7 +774,7 @@ mod tests {
         );
 
         let user_id = user_id!("@username:example.org").as_str();
-        let device_id = device_id!("VALID_DEVICE_ID").as_str();
+        let device_id = device_id!("VALID_DEVICE_ID");
 
         let parse_result = from_json_value::<AnyStateEvent>(member_event_json(user_id));
         assert_matches!(parse_result, Ok(_));